processor/requirements.txt
confluent-kafka==2.3.0
influxdb-client==1.39.0
psycopg2-binary==2.9.9
redis==5.0.1
pandas>=2.2.0
numpy>=1.26.2
python-dotenv==1.0.0
